payment_collections_client_api/requirements.txt
amqp==2.4.0
atomicwrites==1.2.1
attrs==18.2.0
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
Django==2.1.5
django-redis==4.10.0
idna==2.8
kombu==4.2.2.post1
#mesika-libs==1.5.8b0
#mesika-utils==1.0.3
more-itertools==5.0.0
pluggy==0.8.1
psycopg2-binary==2.7.7
py==1.7.0
pycrypto==2.6.1
pycryptodomex==3.7.3
pytest==4.1.1
pytz==2018.9
redis==3.0.1
requests==2.21.0
six==1.12.0
Unidecode==1.0.23
urllib3==1.24.1
uWSGI==2.0.17.1
vine==1.2.0
